This commit is a bug test case for the asclite scorer when the intersegment gaps are expected to be added to the scoring output but they are not.
When the isgbug-work.ctm file is scored, inter-segment-gaps are
produced but for the isgbug-fail.ctm file, the inter-segment-gaps are
not produced.

This feels like a condition problem somewhere but I'm not sure were.
To run the test, cd into src/asclite/testfiles and run:

     perl ./ascliteTest.pl -s todebug

While the test runs successfully, it should not.  The
ascliteTestOutDir/isgbug-fail.sgml should have inter-segment-gaps like
ascliteTestOutDir/isgbug-work.sgml.